requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
fastapi==0.115.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
ollama==0.2.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-telegram-bot==21.4
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2